% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{as.dlist}
\alias{as.dlist}
\title{Creates a dlist from the input.}
\usage{
as.dlist(items, nparts = NULL)
}
\arguments{
\item{items}{The object to convert to a DList}

\item{nparts}{The number of partitions for the resulting DList to have.}
}
\value{
A DList converted from the input.
Note that a list of partitions (resulting from the use of parts()) may
be used with as.dlist. This will recombine those partitions into a single
DObject.
}
\description{
Creates a dlist from the input.
}
\examples{
\dontrun{
a <- as.dlist(list(1,2,3,4)) # A DList with elements 1 to 4.
b <- as.dlist(parts(a,c(3,4))) # A new DList with only 2 partitions, which
were partitions 3 and 4 of 'a'.
}
}
\seealso{
\code{\link{dlist}}
}

