% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dds.R, R/pobject.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,ParallelDDS,list-method}
\title{Combines a list of partitions into a single distributed
object. (can be implemented by a frontend wrapper without actually
combining data in storage).}
\usage{
combine(driver, items)

\S4method{combine}{ParallelDDS,list}(driver, items)
}
\arguments{
\item{driver}{The driver on which combine is dispatched.}

\item{items}{A list of partitions to combine.}
}
\value{
A new distributed object made form the items list.
}
\description{
Combines a list of partitions into a single distributed
object. (can be implemented by a frontend wrapper without actually
combining data in storage).
}

