% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Test condition}
\usage{
condition(condition, val_true, val_false)
}
\arguments{
\item{condition}{Condition that needs to be checked}

\item{val_true}{Return value when condition is true}

\item{val_false}{Return value when condition is false}
}
\value{
The second or third input parameter based on the condition
}
\description{
Test a condition and returns the according parameter
}
\details{
This function allows to test a condition and get the appropriate
output when TRUE/FALSE with avoiding endless if then else loops
in the code.
}
\examples{
condition(5<10,val_true="5 is smaller than 10",val_false="5 is bigger than 10")

}
\author{
Yves R. Sagaert
}
