% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{darray}
\alias{darray}
\title{Creates a darray with the specified partitioning and data.}
\usage{
darray(nparts = NULL, dim = NULL, psize = NULL, data = 0)
}
\arguments{
\item{nparts}{The number of partitions this DArray should have. If this is not provided or is NULL, then psize and dim must be provided together.}

\item{dim}{The dimensions of the overall DArray to construct. Must be provided together with psize.}

\item{psize}{A 2d-vector indicating the size of each partition. In general, each
dimension of this vector has a value that evenly divides into dim; if not, then the last partition will be smaller. This parameter is provided together with dim.}

\item{data}{(Default value: 0) The value to which each element of the DArray should be initialized to.}
}
\value{
A DArray, with each partition equal to size psize and number of partitions in each dimension equal to nparts, with each element initialized to data.
}
\description{
Creates a darray with the specified partitioning and data.
}
\examples{
\dontrun{
a <- darray(psize=c(3,3),dim=c(9,9),data=5) # A 9 partition (each partition 3x3), 9x9 DArray with each element initialized to 5.
b <- darray(psize=c(3,3),dim=c(9,9)) # Same as 'a', but filled with 0s.
c <- darray(nparts=c(2,3)) # An empty darray with 6 partitions, 2 per column and 3 per row.
}
}

