% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R, R/dds_package.R
\docType{package}
\name{dds}
\alias{dds}
\alias{dds-package}
\title{dds function}
\usage{
dds(x, fun = "", a = TRUE)
}
\arguments{
\item{x}{Data or output of a function which needs description}

\item{fun}{Function that needs to be executed on the data (optional)}

\item{a}{Auditive output required (default = TRUE)
#'}
}
\value{
Textual output and auditive output (when a=TRUE)
}
\description{
Creates textual summary of data or function output.

This package provides descriptive data analysis
in addition to your standard data plotting.
}
\details{
This function converts graphical output to textual and auditive output.
}
\examples{
  \dontrun{
     library(sonify)
     x <- rnorm(100)
     dds(x)
  }

x <- rnorm(100)
dds(x,a=FALSE)

}
\author{
Yves R. Sagaert

Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
