% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{dframe}
\alias{dframe}
\title{Creates a DFrame with the specified partitioning and data.}
\usage{
dframe(nparts = NULL, dim = NULL, psize = NULL, data = 0)
}
\arguments{
\item{nparts}{The number of partitions this DFrame should have. If this is not provided or is NULL, then psize and dim must be provided together.}

\item{dim}{The dimensions of the overall DFrame to construct. Must be provided together with psize.}

\item{psize}{A 2d-vector indicating the size of each partition. In general, each
dimension of this vector has a value that evenly divides into dim; if not, then the last partition will be smaller. This parameter is provided together with dim.}

\item{data}{(Default value: 0) The value to which each element of the DFrame should be intialized to.}
}
\value{
A DFrame, with each partition equal to size psize and number of partition in each dimension equal to nparts, with each element initialized to data.
}
\description{
Creates a DFrame with the specified partitioning and data.
}
\examples{
\dontrun{
a <- dframe(psize=c(3,3),dim=c(9,9),data=5) # A 9 partition (each partition 3x3), 9x9 DFrame with each element initialized to 5.
b <- dframe(psize=c(3,3),dim=c(9,9)) # Same as 'a', but filled with 0s.
c <- dframe(nparts=c(2,3)) # An empty DFrame with 6 partitions, 2 per column and 3 per row.
}
}

