% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{dimnames<-,DObject,list-method}
\alias{dimnames<-,DObject,list-method}
\title{Sets the dimnames for the distributed object.}
\usage{
\S4method{dimnames}{DObject,list}(x) <- value
}
\arguments{
\item{x}{The object to set the dimnames for.}

\item{value}{The list of values, one vector per dimension, of names.}
}
\description{
Sets the dimnames for the distributed object.
}

