% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dds.R
\name{dlapply}
\alias{dlapply}
\title{Like dmapply, but permits only one iterable argument, and the output.type is
always a dlist. This is the distributed version of 'lapply'.}
\usage{
dlapply(X, FUN, ..., nparts = NULL, .unlistEach = FALSE)
}
\arguments{
\item{X}{The vector, matrix, list, data.frame, dlist, darray, or dframe or other iterable object to supply to the function in FUN.}

\item{FUN}{the function to be applied to each element of ‘X’.}

\item{...}{optional arguments to 'FUN'.}
}
\value{
a DList with nparts=nparts
}
\description{
Note that as in lapply vs mapply, the order of arguments (iterable argument and function) supplied to the functions is reversed.
}
\examples{
\dontrun{
a <- dlapply(1:5,function(x) x, nparts=3) # A DList with 3 partitions, which in the aggregate contains the elements 1 through 5.
b <- dlapply(a,function(x) x+3) # Adds 3 to each element of dlist a.
}
}

