% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{dlist}
\alias{dlist}
\title{Creates a dlist with the specified partitioning and data.}
\usage{
dlist(..., nparts = 1L)
}
\arguments{
\item{...}{Values to initialize the DList with (optional).}

\item{nparts}{(Default is 1L) The number of partitions this DList should have.}
}
\value{
A DList containing the data in ..., or an empty DList, partitioned accordingly based on nparts.
}
\description{
Creates a dlist with the specified partitioning and data.
}
\examples{
\dontrun{
a <- dlist(1,2,3,4,nparts=2) # A DList containing 2 partitions, with data 1 to 4.
}
}

