% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\docType{class}
\name{DObject-class}
\alias{DObject-class}
\title{The baseline distributed object class to be extended by each backend driver.
Backends may elect to extend once for all distributed object types ('dlist',
'darray', 'dframe,', etc.) for one per type, depending on needs.}
\description{
The baseline distributed object class to be extended by each backend driver.
Backends may elect to extend once for all distributed object types ('dlist',
'darray', 'dframe,', etc.) for one per type, depending on needs.
}
\section{Slots}{

\describe{
\item{\code{nparts}}{Stores the 2d-partitioning scheme of the distributed object.}

\item{\code{psize}}{Stores, as a 2d-matrix (1d-for dlists) of the size of each
partition.}

\item{\code{dim}}{The dimensions of the distributed object.}

\item{\code{backend}}{A character vector of the name of the backend that created
the object.}

\item{\code{type}}{The distributed object type for this object (e.g,. 'dlist').}
}}

