% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.darray}
\alias{is.darray}
\title{Returns whether the input entity is a DArray}
\usage{
is.darray(x)
}
\arguments{
\item{x}{The input to test to see whether it is a DArray.}
}
\value{
TRUE if x is a DArray, FALSE otherwise
}
\description{
Returns whether the input entity is a DArray
}
\examples{
\dontrun{
is.darray(3) # FALSE
is.darray(darray(psize=c(3,3),dim=c(9,9))) # TRUE
}
}

