% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.dframe}
\alias{is.dframe}
\title{Returns whether the input entity is a DFrame}
\usage{
is.dframe(x)
}
\arguments{
\item{x}{The input to test to see whether it is a DFrame.}
}
\value{
TRUE if x is a DFrame, FALSE otherwise
}
\description{
Returns whether the input entity is a DFrame
}
\examples{
\dontrun{
is.dframe(3) # FALSE
is.dframe(dframe(psize=c(3,3),dim=c(9,9))) # TRUE
}
}

