% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.sparse_darray}
\alias{is.sparse_darray}
\title{Returns whether the input is a sparse_darray}
\usage{
is.sparse_darray(x)
}
\arguments{
\item{x}{input object.}
}
\value{
TRUE if x is a sparse_darray, FALSE otherwise.
}
\description{
Returns whether the input is a sparse_darray
}
\examples{
\dontrun{
is.sparse_darray(3) # FALSE
is.sparse_darray(darray(psize=c(3,3),dim=c(9,9))) # FALSE
is.sparse_darray(darray(npartitions=3,sparse=TRUE)) # TRUE
}
}

