% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{mean,DObject-method}
\alias{mean,DObject-method}
\title{Gets the mean value of the elements within the object.}
\usage{
\S4method{mean}{DObject}(x, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{The distributed object to get the mean of.}

\item{trim}{Not supported yet.}

\item{na.rm}{If TRUE, removes NA values.}

\item{...}{Other args.}
}
\description{
Gets the mean value of the elements within the object.
}

