% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{nparts}
\alias{nparts}
\title{Returns a 2d-vector denoting the number of partitions existing along
each dimension of the DObject, where the vector==c(no_of_partitions_per_column,
no_of_partitions_per_row). Note that for DLists, the value is always
equivalent to c(totalParts(dobj),1).}
\usage{
nparts(dobj)
}
\arguments{
\item{dobj}{The DObject whose 2d partitioning vector to retrieve.}
}
\value{
A 2d-vector containing the number of partitions along each dimension.
}
\description{
Returns a 2d-vector denoting the number of partitions existing along
each dimension of the DObject, where the vector==c(no_of_partitions_per_column,
no_of_partitions_per_row). Note that for DLists, the value is always
equivalent to c(totalParts(dobj),1).
}
\examples{
\dontrun{
a <- darray(psize=c(3,3),dim=c(9,9)) # 9 partitions of 3x3
b <- nparts(a) # returns c(3,3)
}
}
\seealso{
\code{\link{totalParts}}
}

