% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdriver.R
\docType{data}
\name{parallel}
\alias{parallel}
\title{The default parallel driver}
\format{\preformatted{Formal class 'ParallelDDS' [package "dds"] with 4 slots
  ..@ DListClass : chr "ParallelObj"
  ..@ DFrameClass: chr "ParallelObj"
  ..@ DArrayClass: chr "ParallelObj"
  ..@ backendName: chr "parallel"
}}
\usage{
parallel
}
\description{
The default parallel driver
}
\examples{
\dontrun{
useBackend(parallel,executors=4)
}
}
\keyword{datasets}

