% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{psize}
\alias{psize}
\title{Return the 1d (in the case of DLists) or 2d-matrix (DArrays and DFrames)
containing the sizes of each partition of the DObject.
The ith row of the matrix refers to the ith partition of the DObject.
The 1st column refers to the number of rows, and the 2nd is the number of columns.}
\usage{
psize(dobj, index = NULL)
}
\arguments{
\item{dobj}{The DObject to get the partition sizes of}

\item{index}{(Default: NULL), a numeric vector or list containing the indices of the}
}
\value{
A matrix, which is 1d in the case of DLists, and 2d in the case of DArrays and
DFrames, containing the partition sizes of the specified DObjects (if index is NULL), or
the partition sizes of the provided ids of partitions (if index is not NULL).
}
\description{
Return the 1d (in the case of DLists) or 2d-matrix (DArrays and DFrames)
containing the sizes of each partition of the DObject.
The ith row of the matrix refers to the ith partition of the DObject.
The 1st column refers to the number of rows, and the 2nd is the number of columns.
}
\examples{
\dontrun{
a <- darray(psize=c(3,3),dim=c(9,9)) # 9 partitions of 3x3
b <- psize(a) # A 9x2 matrix, with each row containing c(3,3)
}
}
\seealso{
\code{\link{nparts}}, \code{\link{parts}}
partition ids to get the sizes for.
}

