% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{repartition}
\alias{repartition}
\title{Repartitions a DObject.
This function takes two inputs, dobj, and skeleton. These inputs must both be DObjects of the same type and same dimension.
If dobj and skeleton have different internal partitioning, this function will return a new dobject with the same internal data as in dobj but with the partitioning scheme of skeleton.}
\usage{
repartition(dobj, skeleton)
}
\arguments{
\item{dobj}{The DObject whose data is to be preserved, but repartitioned.}

\item{skeleton}{The DObject whose partitioning is to be emulated in the output.}
}
\value{
A new DObject with the data of 'dobj' and the partitioning of 'skeleton'.
}
\description{
Repartitions a DObject.
This function takes two inputs, dobj, and skeleton. These inputs must both be DObjects of the same type and same dimension.
If dobj and skeleton have different internal partitioning, this function will return a new dobject with the same internal data as in dobj but with the partitioning scheme of skeleton.
}
\examples{
\dontrun{
a <- dlist(1,2,3,4,nparts=2)
b <- dmapply(function(x) x, 10:14,nparts=4)
c <- repartition(a,b) # c will have 4 partitions of length 1 each, containing 1 to 4.
}
}

