% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{rownames,DObject-method}
\alias{rownames,DObject-method}
\title{Gets the rownames for the distributed object.}
\usage{
\S4method{rownames}{DObject}(x)
}
\arguments{
\item{x}{The distributed object to get the rownames for.}
}
\description{
Gets the rownames for the distributed object.
}

