% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{rowSums,DObject-method}
\alias{rowSums,DObject-method}
\title{Gets the row sums for a distributed array or data.frame.}
\usage{
\S4method{rowSums}{DObject}(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{The object to get the row sums from.}

\item{na.rm}{If TRUE, will remove NAs.}

\item{dims}{Currently does nothing.}
}
\description{
Gets the row sums for a distributed array or data.frame.
}

