% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dds.R, R/pdriver.R
\docType{methods}
\name{shutdown}
\alias{shutdown}
\alias{shutdown,DDSDriver-method}
\alias{shutdown,ParallelDDS-method}
\title{Called when the backend driver is shutdown.}
\usage{
shutdown(x)

\S4method{shutdown}{DDSDriver}(x)

\S4method{shutdown}{ParallelDDS}(x)
}
\arguments{
\item{x}{The driver object to shutdown.}
}
\description{
Called when the backend driver is shutdown.
}
\section{Methods (by class)}{
\itemize{
\item \code{DDSDriver}: Default backend shutdown message.

\item \code{ParallelDDS}: Shutdown for parallel
}}

