% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{totalParts}
\alias{totalParts}
\title{Returns the total number of partitions of the DObject.
Note that that this is the same result as prod(nparts(dobj))}
\usage{
totalParts(dobj)
}
\arguments{
\item{dobj}{The DObject of which to get the number of partitions.}
}
\value{
The number of partitions the DObject is divided into.
}
\description{
Returns the total number of partitions of the DObject.
Note that that this is the same result as prod(nparts(dobj))
}
\examples{
\dontrun{
a <- darray(psize=c(3,3),dim=c(9,9)) # 9 partitions of 3x3
b <- totalParts(a) # Returns 9
}
}
\seealso{
\code{\link{nparts}}
}

