% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dds.R
\name{useBackend}
\alias{useBackend}
\title{Sets the active backend driver in the DDS environment. The driver object is
responsible for the dispatch of all backend-specific DDS operations, such as
dmapply, collect, and parts. Besides allowing the backend driver to be changed,
initialization parameters may be passed into the ellipsis (...) part of the
function arguments.}
\usage{
useBackend(driver, ...)
}
\arguments{
\item{driver}{The driver object to load as the active backend for DDS. This object should extend class 'DDSDriver', and the S4 methods for do_dmapply, do_collect, and get_parts should be defined for the class of the driver object.}

\item{...}{Additional parameters to pass to the initialization function of this driver.
After successfully setting the backend to a new backend with useBackend(), all subsequent dmapply, collect, and parts operations will dispatch on that driver object's class. Note that DObjects created with a different backend prior to switching will be incompatible with these backend-specific functions of the new driver.}
}
\description{
The default loaded driver in DDS is parallel.
}
\examples{
\dontrun{
useBackend(parallel,inst=2)
library(distributedR.dds); useBackend(distributedR)
}
}

