% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R, R/dds_package.R
\docType{package}
\name{dds}
\alias{dds}
\alias{dds-package}
\title{dds function}
\usage{
dds(x, fun = "", a = TRUE)
}
\arguments{
\item{x}{Data or output of a function which needs description}

\item{fun}{Function that needs to be executed on the data (optional).
Argument fun can be: c("ts","res", "acf"). See examples for
individual use of fun (x is changing alongside)}

\item{a}{Auditive output required (default = TRUE)
#'}
}
\value{
Textual output and auditive output (when a=TRUE)
}
\description{
Creates textual summary of data or function output.

This package provides descriptive data analysis
in addition to your standard data plotting.
}
\details{
This function converts graphical output to textual and auditive output.
}
\examples{
  \dontrun{
     library(sonify)
     x <- rnorm(100)
     dds(x)
  }

# Any data
set.seed(1010)
x <- rnorm(100)
dds(x,a=FALSE)

# Time series data
set.seed(1010)
x <- ts(rnorm(100),start=c(2020,1),freq=12)
dds(x,a=FALSE,fun="ts")

# Residual analysis
library(forecast)
set.seed(1010)
x <- ts(rnorm(100),start=c(2020,1),freq=12)
dds(naive(x),a=FALSE,fun="res")
}
\author{
Yves R. Sagaert

Yves R. Sagaert \email{yves.r.sagaert@gmail.com}
}
