% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.darray}
\alias{is.DArray}
\alias{is.darray}
\title{Returns whether the input is a darray}
\usage{
is.darray(x)

is.DArray(x)
}
\arguments{
\item{x}{input object.}
}
\value{
TRUE if x is a darray, FALSE otherwise.
}
\description{
Returns whether the input is a darray
}
\examples{
\dontrun{
is.darray(3) # FALSE
is.darray(darray(psize=c(3,3),dim=c(9,9))) # TRUE
}
}

