% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.dlist}
\alias{is.dlist}
\title{Returns whether the input entity is a DList}
\usage{
is.dlist(x)
}
\arguments{
\item{x}{The input to test to see whether it is a DList.}
}
\value{
TRUE if x is a DList, FALSE otherwise
}
\description{
Returns whether the input entity is a DList
}
\examples{
\dontrun{
is.dlist(3) #FALSE
is.dlist(dlist(1,2,3,nparts=3)) #TRUE
}
}

